# variable-argument functions