--- a/compiler/ast/ident_list.go
+++ b/compiler/ast/ident_list.go
@@ -1,6 +1,7 @@
 package ast
 
 import (
+	"fmt"
 	"strings"
 
 	"github.com/d5/tengo/compiler/source"
@@ -50,6 +51,10 @@ func (n *IdentList) NumFields() int {
 }
 
 func (n *IdentList) String() string {
+	if n.VarArgs {
+		return fmt.Sprintf("(...%s)", n.List[0].Name)
+	}
+
 	var list []string
 	for _, e := range n.List {
 		list = append(list, e.String())