This is just an initial implementation of variable-argument functions. I'm expecting to need to make changes, but I wanted to demonstrate the initial approach and get feedback before going further.

With this implementation, there are two types of functions supported: regular (`func(a,b,c) {}`) and variadic (`func(...args) {}`). Currently, if a function is variadic, it must have exactly one argument. The VM is responsible for consolidating passed parameters when it processes an `OpCall` instruction on a `CompiledFunction` or `Closure` that is variadic. It does this by consolidating the arguments on the stack into an array, which is then stored on the stack in place of the original first argument. If no arguments are passed to a variadic function call, an empty array is added to the stack as the first argument.