@d5 2019-04-24
@mccolljr I will invite you to a Slack channel if you send me an email: me@daniel.gs