@earncef 2019-04-24
The other concern I have is of performance. I modified the two fib tail call functions in `cmd/bench`. The variadic implementation is quite slow in both cases. I'm assuming it adds a significant overhead when the arguments are consolidated into an array and pushed to the stack. 

The modifications and results:

```
fib := func(...x) {
	if x[0] == 0 {
		return 0 + x[1]
	} else if x[0] == 1 {
		return 1 + x[1]
	}

	return fib(x[0]-1, fib(x[0]-2, x[1]))
}
```

```
fib := func(...x) {
	if x[0] == 0 {
		return x[1]
	} else if x[0] == 1 {
		return x[2]
	}

	return fib(x[0]-1, x[2], x[1]+x[2])
}
```

Results:

BEFORE:
```
-------------------------------------
fibonacci(35) (tail-call #1)
-------------------------------------
Result:  9227465
Go:      79.25467ms
Parser:  23.245µs
Compile: 99.047µs
VM:      3.026240441s
-------------------------------------
fibonacci(35) (tail-call #2)
-------------------------------------
Result:  9227465
Go:      442ns
Parser:  28.094µs
Compile: 85.925µs
VM:      16.416µs
```

AFTER:

```
-------------------------------------
fibonacci(35) (tail-call #1)
-------------------------------------
Result:  9227465
Go:      74.706679ms
Parser:  59.182µs
Compile: 106.193µs
VM:      8.298983466s
-------------------------------------
fibonacci(35) (tail-call #2)
-------------------------------------
Result:  9227465
Go:      282ns
Parser:  36.099µs
Compile: 107.335µs
VM:      42.433µs
```