@mccolljr 2019-04-24
@earncef that performance hit is pretty big... but I have to wonder how likely it is that tail-call recursion is going to be used in conjunction with variadic functions. I tried to keep changes to the core structure of the language to a minimum, which meant that the only place I could get information about the callee was in at the vm level, when an `OpCall` instruction is processed. That's why the process of consolidating arguments into an array for variadic functions happens at the vm level. I'm open to suggestions or a nudge in the right direction but I'm not sure what the best way to improve the situation is. Maybe there's some special handling that can be done in the tail call check to update indexes in the argument array rather than pushing the arguments back on the stack? I'l look into it when I have some time later today, but advice would be appreciated.